chore: lower linecheck gate from 2000 to 1500 lines
Splits service_tests.rs (1770 lines) by extracting prompt-validation and
I/O-failure tests into service_coverage_tests.rs. All .rs files now under
1500 lines.

Closes #934
