+++ .changeset/linecheck-1500.md
@@ -1,0 +1,5 @@
+---
+"moadim": patch
+---
+
+chore: lower linecheck gate from 2000 to 1500 lines
