+++ .githooks/pre-push
@@ -84,9 +84,9 @@ if ! command -v linecheck >/dev/null 2>&1; then
   exit 1
 fi
 
-echo "Checking .rs files do not exceed 2000 lines..."
+echo "Checking .rs files do not exceed 1500 lines..."
 # shellcheck disable=SC2046
-linecheck --max-lines 2000 $(find src -name '*.rs')
+linecheck --max-lines 1500 $(find src -name '*.rs')
 echo "Line-count check passed."
 
 if [ "${SKIP_CHANGELOG:-0}" != "1" ]; then
